backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-dotenv
google-generativeai
pydantic
flake8
python-jose[cryptography]
werkzeug